// src/dto/lov-item.dto.ts
/**
 * Created by sandip on 10/05/17.
 */


export class SecurityQuestionLov {
    code: string;
    descr: string;
}

export class CountryLov {
    code: string;
    descr: string;
    isdCode: string;
}

export class StateLov {
    code: string;
    descr: string;
}

export class CityLov {
    code: string;
    descr: string;
}

export class NameSuffixLov {
    code: string;
    descr: string;
}